Changed DataBuffer to be more 32/64 bit agnostic. DataBuffers will be the same size between ILP32, LP64, and LLP64 (note that ILP64 isn't supported). Changed message to have a maximum message length of unsigned short (65535) which should be enough for normal messages (files are a different story). The message length might be configurable sometime in the future if it becomes an issue.

+++ include/Message.hpp
@@ -60,8 +60,8 @@ class Message
     private:
 
         bool hasReadLength; // if the length has been determined
-        size_t bytesLeft;   // the bytes left to read
-        size_t length;      // the length of the message to read
+        unsigned short bytesLeft;   // the bytes left to read
+        unsigned short length;      // the length of the message to read
         DataBuffer buffer;  // the databuffer
 };
